add abstract class Component, refactor trip-point, trip-point-edit and filter components
